refactor: extract DOM posting attempt policy
