+++ src/background/platform-poster.ts
@@ -14,7 +14,6 @@ import { t } from '../utils/i18n';
 import {
   closeTabSafely,
   openOrFocusTab,
-  type OpenOrFocusTabOptions,
 } from './tab-management';
 import {
   buildExpectedUrlsForVerification,
@@ -25,7 +24,6 @@ import {
   isVerifySupported,
   resolveComposeUrlForMedia,
   runVerify,
-  shouldForceInlineThreadPreviewForeground,
   shouldUseInlineThread,
   tryApiPath,
 } from './platform-strategies';
@@ -48,19 +46,18 @@ import {
 import type { OpenedTabRegistry } from './opened-tab-registry';
 import { retryTransientTabAction } from './tab-action-retry';
 import type { VerifyExpectation } from '../utils/post-verify';
+import {
+  buildDomPostAttempts,
+  type DomPostAttempt,
+  resolvePreSubmitLoadOptions,
+  shouldOpenActive,
+  shouldRetryPostAttempt,
+  shouldReuseExistingTabForAttempt,
+} from './dom-attempt-policy';
 
 const CHUNK_INTERVAL_MS = 2000;
 
 type Visibility = 'public' | 'unlisted' | 'private' | 'direct';
-
-export interface DomPostAttempt {
-  label: string;
-  skipApi?: boolean;
-  forceActive?: boolean;
-  reuseExistingTab?: boolean;
-  loadRetries?: number;
-  delayBeforeMs?: number;
-}
 
 export interface PlatformPosterOptions {
   openedTabs: Pick<OpenedTabRegistry, 'record' | 'forget'>;
@@ -567,13 +564,6 @@ export function resolveApiPostOutcome(
   });
 }
 
-export function shouldRetryPostAttempt(
-  autoPost: boolean,
-  flow?: Pick<PostFlowTrace, 'submitReached'>,
-): boolean {
-  return !autoPost || flow?.submitReached !== true;
-}
-
 export function isAmbiguousPostDispatchError(err: unknown): boolean {
   const message = err instanceof Error ? err.message : String(err);
   return (
@@ -591,74 +581,6 @@ export function getComposeUrlForMedia(
   images?: readonly ImageAttachment[],
 ): string {
   return resolveComposeUrlForMedia(adapter.id, adapter.getComposeUrl(text), images);
-}
-
-export function resolvePreSubmitLoadOptions(
-  adapter: Pick<PlatformAdapter, 'preSubmitLoad'>,
-): OpenOrFocusTabOptions | undefined {
-  const policy = adapter.preSubmitLoad;
-  if (!policy) return undefined;
-  return {
-    loadRetries: policy.retryCount,
-    relaxedComposeUrlReady: policy.urlReady === 'same-origin-path',
-  };
-}
-
-export function shouldOpenActive(
-  adapter: PlatformAdapter,
-  dryRun: boolean,
-  textChunks?: string[],
-  autoPost = !dryRun,
-  forceForeground = false,
-): boolean {
-  if (forceForeground) return true;
-  if (autoPost) return true;
-  const forceForegroundForThreadPreview =
-    shouldForceInlineThreadPreviewForeground(adapter.id, dryRun, textChunks);
-  return adapter.requiresForegroundTab === true || forceForegroundForThreadPreview;
-}
-
-export function buildDomPostAttempts(
-  adapter: PlatformAdapter,
-  autoPost: boolean,
-  forceForeground = false,
-): DomPostAttempt[] {
-  const dryRun = !autoPost;
-  const attempts: DomPostAttempt[] = [
-    forceForeground ? { label: 'default', forceActive: true } : { label: 'default' },
-    {
-      label: 'fresh foreground compose',
-      skipApi: true,
-      forceActive: true,
-      reuseExistingTab: false,
-      loadRetries: 1,
-      delayBeforeMs: dryRun ? 250 : 750,
-    },
-  ];
-
-  if (!adapter.requiresForegroundTab) {
-    attempts.push({
-      label: 'fresh foreground compose with reload retry',
-      skipApi: true,
-      forceActive: true,
-      reuseExistingTab: false,
-      loadRetries: 2,
-      delayBeforeMs: 1000,
-    });
-  }
-
-  return attempts;
-}
-
-export function shouldReuseExistingTabForAttempt(
-  adapter: Pick<PlatformAdapter, 'requiresForegroundTab'>,
-  autoPost: boolean,
-  attempt: Pick<DomPostAttempt, 'reuseExistingTab'> = {},
-  forceForeground = false,
-): boolean {
-  if (typeof attempt.reuseExistingTab === 'boolean') return attempt.reuseExistingTab;
-  const dryRun = !autoPost;
-  return dryRun && adapter.requiresForegroundTab !== true && !forceForeground;
 }
 
 async function attachVerifyResult(
